+++ packages/contracts/src/framework/plugin/setup/PluginSetup.sol
@@ -14,6 +14,12 @@ import {IPluginSetup} from "./IPluginSetup.sol";
 /// @author Aragon Association - 2022-2023
 /// @notice An abstract contract that developers have to inherit from to write the setup of a plugin.
 abstract contract PluginSetup is ERC165, IPluginSetup {
+    address private immutable implementation;
+
+    constructor(address _implementation) {
+        implementation = _implementation;
+    }
+
     /// @inheritdoc IPluginSetup
     function prepareUpdate(
         address _dao,
@@ -46,5 +52,7 @@ abstract contract PluginSetup is ERC165, IPluginSetup {
     }
 
     /// @inheritdoc IPluginSetup
-    function getImplementationAddress() external view virtual returns (address) {}
+    function getImplementationAddress() external view returns (address) {
+        return implementation;
+    }
 }